app/orders/serializers.py
from core.models import Order
from rest_framework import serializers


class OrderSerializer(serializers.ModelSerializer):
    class Meta:
        model = Order
        fields = '__all__'

        read_only_fields = ('id', 'created', 'updated', 'user')
